refactor: update folder strcuture for home
