+++ src/pages/Home/components/SignUp/index.tsx
@@ -11,8 +11,8 @@ import { gql } from "apollo-boost";
 import { useFormik, Formik } from "formik";
 import React, { useContext, useEffect, useState } from "react";
 
-import FullscreenLoadingContext from "../../../context/Loading";
-import HomeInput from "../components/Input";
+import FullscreenLoadingContext from "../../../../context/Loading";
+import HomeInput from "../Input";
 
 import Styled from "./styled";
