+++ src/pages/Home/components/SignUp/styled.ts
